+++ front/lib/actions/mcp_internal_actions/index.ts
@@ -1,7 +1,7 @@
 import { MCPServerNotFoundError } from "@app/lib/actions/mcp_errors";
 import { getInternalMCPServerNameAndWorkspaceId } from "@app/lib/actions/mcp_internal_actions/constants";
 import type { InMemoryWithAuthTransport } from "@app/lib/actions/mcp_internal_actions/in_memory_with_auth_transport";
-import { getInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/servers";
+import { getInternalMCPServer } from "@app/lib/api/actions/servers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
 import type { Authenticator } from "@app/lib/auth";
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
