+++ front/lib/api/actions/servers/jira/jira_utils.ts
@@ -2,8 +2,8 @@ import type {
   SearchFilter,
   SearchFilterField,
   SortDirection,
-} from "@app/lib/actions/mcp_internal_actions/servers/jira/types";
-import { FIELD_MAPPINGS } from "@app/lib/actions/mcp_internal_actions/servers/jira/types";
+} from "@app/lib/api/actions/servers/jira/types";
+import { FIELD_MAPPINGS } from "@app/lib/api/actions/servers/jira/types";
 import { markdownToAdf } from "marklassian";
 
 // Helper function to escape JQL values that contain spaces or special characters
