+++ front/lib/api/actions/servers/jira/metadata.ts
@@ -6,7 +6,7 @@ import {
   JiraSearchFilterSchema,
   JiraSortSchema,
   SEARCH_USERS_MAX_RESULTS,
-} from "@app/lib/actions/mcp_internal_actions/servers/jira/types";
+} from "@app/lib/api/actions/servers/jira/types";
 import type { ServerMetadata } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { createToolsRecord } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import type { JSONSchema7 as JSONSchema } from "json-schema";
