+++ front/lib/api/actions/servers/microsoft_drive/tools/index.ts
@@ -6,7 +6,7 @@ import {
   searchMicrosoftDriveItems,
   validateDocumentXml,
   validateZipFile,
-} from "@app/lib/actions/mcp_internal_actions/servers/microsoft/utils";
+} from "@app/lib/api/actions/servers/microsoft/utils";
 import type { ToolHandlers } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import {
