+++ front/lib/api/actions/servers/microsoft_teams/tools/index.ts
@@ -4,8 +4,8 @@ import type {
   TeamsChat,
   TeamsMessage,
   TeamsUser,
-} from "@app/lib/actions/mcp_internal_actions/servers/microsoft/utils";
-import { getGraphClient } from "@app/lib/actions/mcp_internal_actions/servers/microsoft/utils";
+} from "@app/lib/api/actions/servers/microsoft/utils";
+import { getGraphClient } from "@app/lib/api/actions/servers/microsoft/utils";
 import type { ToolHandlers } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { MICROSOFT_TEAMS_TOOLS_METADATA } from "@app/lib/api/actions/servers/microsoft_teams/metadata";
