+++ front/lib/api/actions/servers/slack_bot/tools/index.ts
@@ -5,7 +5,7 @@ import {
   executePostMessage,
   executeSearchUser,
   getSlackClient,
-} from "@app/lib/actions/mcp_internal_actions/servers/slack/helpers";
+} from "@app/lib/api/actions/servers/slack/helpers";
 import type {
   ToolDefinition,
   ToolHandlers,
