+++ front/lib/api/actions/servers/sound_studio/tools/index.ts
@@ -2,7 +2,7 @@ import { MCPError } from "@app/lib/actions/mcp_errors";
 import {
   getElevenLabsClient,
   streamToBase64,
-} from "@app/lib/actions/mcp_internal_actions/servers/elevenlabs/utils";
+} from "@app/lib/api/actions/servers/elevenlabs/utils";
 import type { ToolHandlers } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { SOUND_STUDIO_TOOLS_METADATA } from "@app/lib/api/actions/servers/sound_studio/metadata";
