+++ front/lib/api/actions/servers/speech_generator/tools/index.ts
@@ -3,7 +3,7 @@ import {
   getElevenLabsClient,
   resolveDefaultVoiceId,
   streamToBase64,
-} from "@app/lib/actions/mcp_internal_actions/servers/elevenlabs/utils";
+} from "@app/lib/api/actions/servers/elevenlabs/utils";
 import type { ToolHandlers } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { buildTools } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { SPEECH_GENERATOR_TOOLS_METADATA } from "@app/lib/api/actions/servers/speech_generator/metadata";
