+++ front/lib/api/actions/servers/tables_query/dialect_instructions.ts
